Vehicle list page added
